Rich text: extract splitValue to separate file
