+++ packages/block-editor/src/components/rich-text/index.js
@@ -15,8 +15,6 @@ import {
 	__unstableUseRichText as useRichText,
 	__unstableCreateElement,
 	isEmpty,
-	split,
-	toHTMLString,
 	isCollapsed,
 	removeFormat,
 } from '@wordpress/rich-text';
@@ -141,87 +139,6 @@ function RichTextWrapper(
 			selectionChange( clientId, identifier, start, end );
 		},
 		[ clientId, identifier ]
-	);
-
-	/**
-	 * Signals to the RichText owner that the block can be replaced with two
-	 * blocks as a result of splitting the block by pressing enter, or with
-	 * blocks as a result of splitting the block by pasting block content in the
-	 * instance.
-	 *
-	 * @param  {Object} record       The rich text value to split.
-	 * @param  {Array}  pastedBlocks The pasted blocks to insert, if any.
-	 */
-	const splitValue = useCallback(
-		( record, pastedBlocks = [] ) => {
-			if ( ! onReplace || ! onSplit ) {
-				return;
-			}
-
-			const blocks = [];
-			const [ before, after ] = split( record );
-			const hasPastedBlocks = pastedBlocks.length > 0;
-			let lastPastedBlockIndex = -1;
-
-			// Consider the after value to be the original it is not empty and
-			// the before value *is* empty.
-			const isAfterOriginal = isEmpty( before ) && ! isEmpty( after );
-
-			// Create a block with the content before the caret if there's no pasted
-			// blocks, or if there are pasted blocks and the value is not empty.
-			// We do not want a leading empty block on paste, but we do if split
-			// with e.g. the enter key.
-			if ( ! hasPastedBlocks || ! isEmpty( before ) ) {
-				blocks.push(
-					onSplit(
-						toHTMLString( {
-							value: before,
-							multilineTag,
-						} ),
-						! isAfterOriginal
-					)
-				);
-				lastPastedBlockIndex += 1;
-			}
-
-			if ( hasPastedBlocks ) {
-				blocks.push( ...pastedBlocks );
-				lastPastedBlockIndex += pastedBlocks.length;
-			} else if ( onSplitMiddle ) {
-				blocks.push( onSplitMiddle() );
-			}
-
-			// If there's pasted blocks, append a block with non empty content
-			/// after the caret. Otherwise, do append an empty block if there
-			// is no `onSplitMiddle` prop, but if there is and the content is
-			// empty, the middle block is enough to set focus in.
-			if (
-				hasPastedBlocks
-					? ! isEmpty( after )
-					: ! onSplitMiddle || ! isEmpty( after )
-			) {
-				blocks.push(
-					onSplit(
-						toHTMLString( {
-							value: after,
-							multilineTag,
-						} ),
-						isAfterOriginal
-					)
-				);
-			}
-
-			// If there are pasted blocks, set the selection to the last one.
-			// Otherwise, set the selection to the second block.
-			const indexToSelect = hasPastedBlocks ? lastPastedBlockIndex : 1;
-
-			// If there are pasted blocks, move the caret to the end of the selected block
-			// Otherwise, retain the default value.
-			const initialPosition = hasPastedBlocks ? -1 : 0;
-
-			onReplace( blocks, indexToSelect, initialPosition );
-		},
-		[ onReplace, onSplit, multilineTag, onSplitMiddle ]
 	);
 
 	const {
@@ -388,7 +305,7 @@ function RichTextWrapper(
 						tagName,
 						onReplace,
 						onSplit,
-						splitValue,
+						onSplitMiddle,
 						__unstableEmbedURLOnPaste,
 						multilineTag,
 						preserveWhiteSpace,
@@ -399,10 +316,10 @@ function RichTextWrapper(
 						value,
 						onReplace,
 						onSplit,
-						multiline,
+						onSplitMiddle,
+						multilineTag,
 						onChange,
 						disableLineBreaks,
-						splitValue,
 						onSplitAtEnd,
 					} ),
 					anchorRef,
